Refactoring: VirtualDisplay<T>, update logic moved
Created a VirtualDisplay<T> struct that represents the display of
CHIP-8 emulator.
VirtualDisplay<T> implements EventHandler, a trait needed by ggez.
The previous implementor was deleted, and the logic that belonged to CPU
was moved back to its run() function from the update() implementation of
the trait.

Deleted some misleading comments.